Move CTABanner, Nav, and Footer to their own files.
