+++ src/layouts/index.jsx
@@ -1,326 +1,23 @@
 // @flow
 
 import React, { useEffect, type Node } from 'react';
-import { StaticQuery, graphql, Link } from 'gatsby';
-import { I18nProvider, withI18n, Trans } from '@lingui/react';
+import { StaticQuery, graphql } from 'gatsby';
+import { I18nProvider, withI18n } from '@lingui/react';
 import { Helmet } from 'react-helmet';
-import sample from 'lodash/sample';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faEnvelope } from '@fortawesome/free-solid-svg-icons';
-import {
-  faTwitter,
-  faFacebook,
-  faLinkedin,
-  faAngellist,
-  faYoutube
-} from '@fortawesome/free-brands-svg-icons';
 import 'typeface-roboto-slab'; // eslint-disable-line import/extensions
 import 'typeface-work-sans'; // eslint-disable-line import/extensions
 import 'katex/dist/katex.min.css';
 import 'prism-themes/themes/prism-ghcolors.css';
 
-import {
-  callToActionExperiments,
-  Title,
-  name,
-  appUrl,
-  loadScript,
-  targetBlank,
-  animateTablet,
-  trackSignup,
-  isBrowser
-} from './utils';
+import { Title, name, loadScript, animateTablet } from './utils';
 import { catalogs, langFromPath, langPrefix, deprefix } from '../i18n-config';
 
 import '../assets/scss/page.scss';
 
-import Modal from '../components/Modal';
-import logoDefault from '../img/logo_black.png';
-import logoInverse from '../img/logo_white.png';
-import NewsletterForm from '../components/NewsletterForm';
 import PublicityBanner from '../components/PublicityBanner';
 import Loader from '../components/Loader';
-
-const Logo = (props: { prefix: string, inverse: boolean }) => (
-  <Link href to={`${props.prefix}/#start`} className="navbar-brand">
-    <img className="logo-dark" src={logoDefault} width={100} height={40} alt={name} />
-    {props.inverse && (
-      <img className="logo-light" src={logoInverse} width={100} height={40} alt={name} />
-    )}
-  </Link>
-);
-
-const navbarLinks = () => [
-  [<Trans>Features</Trans>, 'features'],
-  [<Trans>Pricing</Trans>, 'pricing'],
-  [<Trans>Partners</Trans>, 'partners'],
-  [<Trans>Help</Trans>, 'help'],
-  [<Trans>Blog</Trans>, 'blog']
-];
-
-const Nav = (props: LayoutProps) => (
-  <nav className="navbar navbar-expand-lg navbar-light navbar-stick-dark" data-navbar="sticky">
-    <div className="container flex-nowrap">
-      <div className="navbar-left">
-        <button className="navbar-toggler">&#9776;</button>
-        <Logo {...props} inverse />
-      </div>
-
-      <section className="navbar-mobile">
-        <h6 className="d-sm-none">Ledgy</h6>
-        <nav className="nav nav-navbar ml-auto">
-          {navbarLinks().map(([label, to]) => (
-            <Link className="nav-link" key={to} href to={`${props.prefix}/${to}/`}>
-              {label}
-            </Link>
-          ))}
-        </nav>
-
-        <span className="navbar-divider" />
-      </section>
-
-      <div className="navbar-right">
-        <a
-          className="btn btn-round btn-outline-light mx-2"
-          href={`${appUrl}/login`}
-          onClick={() => trackSignup('login')}
-        >
-          <Trans>Log In</Trans>
-        </a>
-        <a
-          className="btn btn-round btn-light ml-lg-0 mr-2"
-          href={`${appUrl}/signup`}
-          onClick={() => trackSignup('clickSignup')}
-        >
-          <Trans>Sign up</Trans>
-        </a>
-      </div>
-    </div>
-  </nav>
-);
-
-const CTABanner = () => {
-  const experiment = isBrowser ? sample(callToActionExperiments) : callToActionExperiments[0];
-  return (
-    <section className="section bg-pale-secondary">
-      <div className="container py-md-4">
-        <div className="row m-0 w-100 justify-content-center align-items-center">
-          <h4 className="m-3 text-center">{experiment.title}</h4>
-          <a
-            className="cta-button m-3 btn btn-lg btn-round btn-primary align-self-center"
-            href={`${appUrl}/signup`}
-            onClick={() => {
-              if (window.ga) window.ga('set', 'dimension2', experiment.name);
-              trackSignup('clickSignup');
-            }}
-          >
-            <Trans>Get started</Trans>
-          </a>
-        </div>
-      </div>
-    </section>
-  );
-};
-
-const FooterCol = ({
-  order,
-  children,
-  wide
-}: {
-  order: number,
-  children: Node,
-  wide?: boolean
-}) => <div className={`col-${wide ? 12 : 6} col-md-2 pl-6 order-md-${order}`}>{children}</div>;
-
-const FooterColBody = ({ title, children }: { title: Node, children: Array<Node> }) => (
-  <>
-    <h6 className="mb-4 mt-1">
-      <strong>{title}</strong>
-    </h6>
-    <div className="nav flex-column">{children}</div>
-  </>
-);
-const companyLinks = [
-  [<Trans>About us</Trans>, 'about-us'],
-  [<Trans>Blog</Trans>, 'blog'],
-  [<Trans>Partners</Trans>, 'partners'],
-  [<Trans>Webinars</Trans>, 'webinars'],
-  [<Trans>Security</Trans>, 'security'],
-  [<Trans>Privacy</Trans>, 'privacy'],
-  [<Trans>Career</Trans>, 'jobs'],
-  [<Trans>Contact & Imprint</Trans>, 'contact']
-];
-const helpLinks = [
-  [<Trans>Help Center</Trans>, 'help'],
-  [<Trans>Getting Started</Trans>, 'help/getting-started'],
-  [<Trans>Glossary</Trans>, 'glossary'],
-  [<Trans>ESOP Templates</Trans>, 'help/employee-participation-guide'],
-  [<Trans>Round Calculator</Trans>, 'calculator']
-];
-const blogLinks = [
-  [<Trans>Option Pools</Trans>, 'blog/pre-and-post-money-option-pools'],
-  [<Trans>Convertible Loans</Trans>, 'blog/convertible-loans'],
-  [<Trans>KPIs & Reports</Trans>, 'updates/kpis-and-reports']
-];
-const productLinks = [
-  [<Trans>Features</Trans>, 'features'],
-  [<Trans>Cap Table</Trans>, 'features/captable'],
-  [<Trans>Modeling</Trans>, 'features/modeling'],
-  [<Trans>Employee Incentives</Trans>, 'features/esop'],
-  [<Trans>Due Diligence</Trans>, 'features/collaboration'],
-  [<Trans>Investors</Trans>, 'features/investors'],
-  [<Trans>Tokenization</Trans>, 'features/tokenization'],
-  [<Trans>Pricing</Trans>, 'pricing']
-];
-const legalLinks = [
-  [<Trans>Terms of Service</Trans>, 'legal/terms'],
-  [<Trans>Privacy Policy</Trans>, 'legal/privacy-policy'],
-  [<Trans>Cookie Policy</Trans>, 'legal/cookie-policy'],
-  [<Trans>GDPR</Trans>, 'legal/gdpr']
-];
-
-const Footer = (props: LayoutProps) => (
-  <div>
-    <CTABanner {...props} />
-    <footer className="footer pb-9 pt-7 py-md-7 px-4 text-white">
-      <div className="row gap-y justify-content-md-center">
-        <FooterCol order={2}>
-          <FooterColBody title={<Trans>Company</Trans>}>
-            {companyLinks.map(([label, link]) => (
-              <Link className="nav-link" href to={`${props.prefix}/${link}/`} key={link}>
-                {label}
-              </Link>
-            ))}
-          </FooterColBody>
-        </FooterCol>
-        <FooterCol order={4}>
-          <FooterColBody title={<Trans>Resources</Trans>}>
-            {helpLinks.map(([label, link]) => (
-              <Link className="nav-link" href to={`${props.prefix}/${link}/`} key={link}>
-                {label}
-              </Link>
-            ))}
-          </FooterColBody>
-          <FooterColBody title={<Trans>Blog</Trans>}>
-            {blogLinks.map(([label, link]) => (
-              <Link className="nav-link" href to={`${props.prefix}/${link}/`} key={link}>
-                {label}
-              </Link>
-            ))}
-          </FooterColBody>
-        </FooterCol>
-        <FooterCol order={3}>
-          <FooterColBody title={<Trans>Product</Trans>}>
-            {productLinks.map(([label, link]) => (
-              <Link className="nav-link" href to={`${props.prefix}/${link}/`} key={link}>
-                {label}
-              </Link>
-            ))}
-          </FooterColBody>
-        </FooterCol>
-        <FooterCol order={5}>
-          <FooterColBody title={<Trans>Legal</Trans>}>
-            {legalLinks.map(([label, link]) => (
-              <Link className="nav-link" href to={`${props.prefix}/${link}/`} key={link}>
-                {label}
-              </Link>
-            ))}
-          </FooterColBody>
-        </FooterCol>
-        <FooterCol order={1} wide>
-          <Link href to={`${props.prefix}/#start`} className="navbar-brand">
-            <img className="logo-light" src={logoInverse} width={100} height={40} alt={name} />
-          </Link>
-          <div className="social mt-2">
-            {[
-              ['https://www.youtube.com/channel/UCRkvNQptxoE-ckmTsrme1_w', faYoutube, 'YouTube'],
-              ['https://twitter.com/Ledgy', faTwitter, 'Twitter'],
-              ['https://www.linkedin.com/company/ledgy', faLinkedin, 'LinkedIn'],
-              ['https://www.facebook.com/ledgyCom/', faFacebook, 'Facebook'],
-              ['https://angel.co/ledgy', faAngellist, 'AngelList']
-            ].map(([href, icon, title]) => (
-              <a href={href} key={title} {...targetBlank}>
-                <FontAwesomeIcon icon={icon} title={title} />
-              </a>
-            ))}
-          </div>
-          <div className="newsletter-signup-CTA">
-            <Modal
-              id="newsletter-signup"
-              title={<Trans>Sign up for the Ledgy newsletter</Trans>}
-              buttonText={
-                <>
-                  <FontAwesomeIcon
-                    className="newsletter-icon mr-2"
-                    icon={faEnvelope}
-                    title="Newsletter"
-                  />
-                  <Trans>Newsletter</Trans>
-                </>
-              }
-              hideFooter
-            >
-              <p className="text-dark my-5">
-                <Trans>
-                  Receive important feature updates, exclusive webinar invitations, and
-                  promotions/offers
-                </Trans>
-              </p>
-              <NewsletterForm {...props} />
-            </Modal>
-          </div>
-          <div className="mt-4">
-            <div className="dropdown">
-              <button
-                className="btn btn-round btn-outline-light dropdown-toggle"
-                type="button"
-                id="dropdownMenuButton"
-                data-toggle="dropdown"
-                aria-haspopup="true"
-                aria-expanded="false"
-              >
-                <Trans>Language</Trans>
-              </button>
-              <div
-                className="dropdown-menu"
-                aria-labelledby="dropdownMenuButton"
-                style={{ minWidth: '7rem' }}
-              >
-                <Link className="dropdown-item d-flex" to={deprefix(props.location.pathname)} href>
-                  <span className="mr-1" role="img" aria-label="English">
-                    🍔
-                  </span>
-                  English
-                </Link>
-                <Link
-                  className="dropdown-item d-flex"
-                  to={`/de${deprefix(props.location.pathname)}`}
-                  href
-                >
-                  <span className="mr-1" role="img" aria-label="Deutsch">
-                    🥨
-                  </span>
-                  Deutsch
-                </Link>
-                <Link
-                  className="dropdown-item d-flex"
-                  to={`/fr${deprefix(props.location.pathname)}`}
-                  href
-                >
-                  <span className="mr-1" role="img" aria-label="Français">
-                    🥖
-                  </span>
-                  Français
-                </Link>
-              </div>
-            </div>
-          </div>
-        </FooterCol>
-      </div>
-      <div data-provide="map" />
-    </footer>
-  </div>
-);
+import { Nav } from '../components/Nav';
+import { Footer } from '../components/Footer';
 
 type SiteProps = {
   ...$Exact<Props>,
